requirements.txt
flake8==6.1.0
pytorch-lightning==2.0.3
marshmallow-dataclass==8.5.14
omegaconf==2.3.0
hydra-core==1.3.2
pandas==2.0.2
numpy==1.25.0
torch==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.11.1
tqdm==4.65.0
statsmodels==0.14.0
scikit-learn==1.3.0
aim==3.17.5
